order history page the double changed in order number and the filter page the all functionality worked

+++ lib/view/dashboard/order_history/presentation/pages/display_order_list_item.dart
@@ -47,7 +47,7 @@ class DisplayOrderListItem extends StatelessWidget {
                       style: AppTextStyles.textFieldNameStyle11W500,
                     ),
                     Text(
-                      order.orderId.toString(),
+                      order.orderId.toInt().toString(),
                       style: AppTextStyles.listItemValueStyle12W500,
                     ),
                     Padding(
